Prepare for using py::array_t when exposing values.
